% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IL1_LUSC.R
\docType{data}
\name{IL1_LUSC}
\alias{IL1_LUSC}
\title{IL-1 and squamous cell lung cancer}
\format{
An matrix contains the genetic associations of the exposures and the outcome.
}
\usage{
data(IL1_LUSC)
}
\description{
Data from two genome-wide association studies for protenomics and the internal
lung cancer consortium study. The outcome is the squamous cell lung cancer,
and the exposures are interleukin-1 alpha, interleukin-1 beta, and interleukin-1
receptor antagonist.
}
\references{
\itemize{
  \item{1. Sun et al. (2018) Genomic atlas of the human plasma proteome. \strong{Nature}, 558(7708):73-79}
  \item{2. Ahola-Olli et al. (2017) Genome-wide Association Study Identifies 27 Loci Influencing Concentrations of Circulating Cytokines and Growth Factors. \strong{Am J Hum Genet},100(1):40-50.}
  \item{3. Wang et al. (2014) Rare variants of large effect in BRCA2 and CHEK2 affect risk of lung cancer. \strong{Nat Genet}, 46(7):736-41.}
}
}
\keyword{datasets}
